requirements/base.txt
Django
django-filter
django-environ
djangorestframework
Pillow
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
filemagic
python-magic

